# Optimize compilation of large literal arrays