--- a/benchmark/array_large_literal.yml
+++ b/benchmark/array_large_literal.yml
@@ -1,0 +1,19 @@
+prelude: |
+  def def_array(size)
+    Object.class_eval(<<-END)
+      def array_#{size}
+        x = 1
+        [#{(['x'] * size).join(',')}]
+      end
+    END
+  end
+  def_array(100)
+  def_array(1000)
+  def_array(10000)
+  def_array(100000)
+benchmark:
+  array_100: array_100
+  array_1000: array_1000
+  array_10000: array_10000
+  array_100000: array_100000
+